Clean more stuff up.
